Refactor AnyAllFilter implementation logic DynamicsValue/fake-xrm-easy#200
