+++ src/FakeXrmEasy.Core/Query/FilterExpressionExtensions.cs
@@ -73,45 +73,8 @@ namespace FakeXrmEasy.Query
             if (fe.AnyAllFilterLinkEntity != null)
             {
                 var le = fe.AnyAllFilterLinkEntity;
-                var constantExpression = Expression.Constant(false);
-                
-                //creates and evaluates the inner query expression, and then applies relevant filtering based on JoinOperator
-                var childQueryExpression = new QueryExpression()
-                {
-                    EntityName = le.LinkToEntityName,
-                    Criteria = le.LinkCriteria,
-                    ColumnSet = new ColumnSet(le.LinkToAttributeName),
-                };
-
-                var childQueryElements = childQueryExpression.ToQueryable(context)
-                    .ToList();
-                    
-                    
-                var childIds = childQueryElements.Select(e => e.GetAttributePrimaryKeyIdOrEntityReferenceId(le.LinkToAttributeName))
-                    .ToList();
-             
-                var getAttributeValueExpr = entity.ToAttributeValueExpression(le.LinkFromAttributeName);
-                var containsAttributeExpr = entity.ToContainsAttributeExpression(le.LinkFromAttributeName);
-                
-                if (fe.AnyAllFilterLinkEntity.JoinOperator == JoinOperator.Any)
-                {
-                    BinaryExpression orExpression = Expression.Or(Expression.Constant(false), Expression.Constant(false));
-                    foreach (var id in childIds)
-                    {
-                        var leftHandSideExpression =
-                            typeof(Guid).GetAppropriateCastExpressionBasedOnType(getAttributeValueExpr, id);
-                        var rightHandSideExpression =
-                            TypeCastExpressionExtensions.GetAppropriateTypedValueAndType(id, typeof(Guid));
-                        
-                        var matchExpression = Expression.AndAlso(containsAttributeExpr,
-                            Expression.Equal(leftHandSideExpression, rightHandSideExpression));
-
-                        orExpression = Expression.Or(orExpression, matchExpression);
-                    }
-
-                    expressionsList.Add(orExpression);
-                }
-                
+                var anyAllFilterExpression = le.TranslateAnyAllLinkedEntityToExpression(context, entity);
+                expressionsList.Add(anyAllFilterExpression);
             }
             
             return GenerateMultipleExpressionsWithOperator(fe.FilterOperator, expressionsList);
